Make ConsulConfigurationSource public
`ConsulConfigurationSource` is package private but it is the return value of the public function, `ConsulConfigurationSourceBuilder::build`. This causes problems when using kotlin b/c kotlin cannot assign any variable to type `ConsulConfigurationSource` since it is package private. Since `ConsulConfigurationSourceBuilder::build` returns an instance of `ConsulConfigurationSource`, we should make `ConsulConfigurationSource` a public class.